Improved the modified date logic
I've been making a lot of sweeping changes to the repo that do not
affect the content of the posts and thus should not affect the
modified timestamp. Naively looking up the last commit in git will not
be correct in this case. This change excludes commits that are tagged
with "NOCONTENT" in their commit subject line, and I've rebased to
include this message in all cross-cutting commit messages.

You'd think this could be accomplished with

    git log --follow --grep --invert-grep "NONCONTENT"

but alas, this doesn't work, instead leaving out commits before the
NOCONTENT commits because they do renames
https://stackoverflow.com/a/64468571 . Because of this, I just use a
kludgy

    /usr/bin/env bash -c 'git log --follow -M --pretty="%f %ct" -- \
    <path> \
    | grep -v NONCONTENT' \
    | head -1 \
    | awk '{print $2}'

to filter the commit list.